:truck: move `mock_async_client` fixture, which was used in mocking RichAsyncClient throughout trust registry modules
